feat(realism): properties+confirm+stats — koniec mocków poziomu 1
* Backend
  - parking_lot.owner_customer_id FK → Customer (właściciel po wizardzie /join)
  - POST /api/parking-lots — rejestracja parkingu + cennika w jednej transakcji
  - GET /api/parking-lots/my?customerId= — parkingi zalogowanego właściciela
  - GET /api/stats/overview — publiczne kafelki (Σ miejsc, # parkingów)
  - confirmReservation(): zapisuje rekord Payment (COMPLETED) + provider ref
  - PaymentRepository: poprawka typu PK Long → Integer

* Frontend
  - JoinPage: realny POST /api/parking-lots po kroku 3,
    nadaje rolę owner, prezentuje błędy walidacji
  - Dashboard: ładuje TYLKO parkingi właściciela; pusty stan z CTA do /join
  - HomePage: realne 'Σ miejsc · # parkingów' z /api/stats/overview
  - ReservePage: po createReservation woła confirmReservation —
    status PENDING → CONFIRMED + zapis płatności w bazie

+++ backend/src/main/java/my/parkuj/application/repository/PaymentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import my.parkuj.application.model.Payment;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 
 }
